Combine Similarity Measures

Combine similarity measures for a single dyad

Usage

combine_sim_seq(similarities, weights = NULL)

Arguments

similarities

A list of similarity measures for a single dyad

weights

A numeric vector of weights for each similarity measure (default is equal weights)

Value

A list containing the combined sequence and average similarity

Details

This function combines multiple similarity measures into a single overall similarity score for a single dyad.

Examples

sim1 <- list(sequence = c(0.8, 0.7, 0.9), average = 0.8)
sim2 <- list(sequence = c(0.6, 0.8, 0.7), average = 0.7)
combine_sim_seq(list(sim1, sim2))
#> $sequence
#> [1] 0.70 0.75 0.80
#> 
#> $average
#> [1] 0.75
#>