Calculate Correlation Between Similarity Measures for a Single Dyad

Calculate Correlation Between Similarity Measures for a Single Dyad

Usage

cor_sim_seq(similarities, method = "pearson")

Arguments

similarities

A list of similarity measures for a single dyad

method

The correlation method to use (default is "pearson")

Value

A correlation matrix

Details

This function calculates the correlation between different similarity measures for a single dyad.

Examples

sim1 <- list(sequence = c(0.8, 0.7, 0.9), average = 0.8)
sim2 <- list(sequence = c(0.6, 0.8, 0.7), average = 0.7)
cor_sim_seq(list(sim1, sim2))
#>      [,1] [,2]
#> [1,]  1.0 -0.5
#> [2,] -0.5  1.0