Calculate lexical similarity sequence for a single dyad

This function calculates lexical similarity over a sequence of conversation exchanges within a single dyad.

Usage

lex_sim_seq(conversation, window_size = 3)

Arguments

conversation

A data frame representing the conversation

window_size

An integer specifying the size of the sliding window

Value

A list containing the sequence of similarities and the average similarity

Examples

conversation <- data.frame(
  processed_text = c("Hello world", "World of programming",
                     "Programming is fun", "Fun world of coding")
)
result <- lex_sim_seq(conversation, window_size = 2)
print(result)
#> $sequence
#> [1] 0.3750000 0.4444444
#> 
#> $average
#> [1] 0.4097222
#>